rename sequence module to rollup_data module
